refactor: execute program for cpp and python was modularized for don't repeat code
