+++ src/runner/mod.rs
@@ -1,8 +1,9 @@
 /*
  *  Quick Test: CLI for stress testing in competitive programming
- *  Copyright (C) 2021  Luis Miguel Báez
+ *  Copyright (C) 2021 - Luis Miguel Báez
  *  License: MIT (See the LICENSE file in the repository root directory)
  */
 
 pub mod lang;
 pub mod types;
+pub mod cmd;
